feature/vrg-0001 communication, vm
