+++ app/src/main/AndroidManifest.xml
@@ -24,7 +24,7 @@
         android:theme="@style/Theme.Words504"
         tools:targetApi="31">
         <activity
-            android:name=".recognition.TestVoiceRecognitionActivity"
+            android:name=".recognition.presentation.TestVoiceRecognitionActivity"
             android:exported="false">
 
             <meta-data
