+++ app/src/main/java/ru/easycode/words504/MainActivity.kt
@@ -4,7 +4,7 @@ import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Button
-import ru.easycode.words504.recognition.TestVoiceRecognitionActivity
+import ru.easycode.words504.recognition.presentation.TestVoiceRecognitionActivity
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
