+++ app/src/main/java/ru/easycode/words504/recognition/STTState.kt
