+++ app/src/main/java/ru/easycode/words504/recognition/data/SpeechRecognizerCallback.kt
@@ -1,0 +1,7 @@
+package ru.easycode.words504.recognition.data
+
+interface SpeechRecognizerCallback {
+    fun started()
+    fun finished(result: String)
+    fun error(code: Int)//todo смаппить в домен
+}
