+++ app/src/main/java/ru/easycode/words504/tts/presentation/TTSCommunication.kt
@@ -1,6 +1,5 @@
 package ru.easycode.words504.tts.presentation
 
-import ru.easycode.words504.presentation.Communication
 
 interface TTSCommunication {
