+++ app/src/main/res/layout/activity_test_voice_recognition.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:gravity="center"
     android:orientation="vertical"
-    tools:context=".recognition.TestVoiceRecognitionActivity">
+    tools:context=".recognition.presentation.TestVoiceRecognitionActivity">
 
     <Button
         android:id="@+id/speakButton"
